chore: Minor fixes for deck.gl
