+++ modules/gltools/src/transform/buffer-transform.ts
@@ -3,8 +3,7 @@
 import {assert, Resource} from '@luma.gl/api';
 import {isWebGL2} from '@luma.gl/webgl';
 
-import type {TransformProps} from './transform';
-import type {TransformDrawOptions, TransformRunOptions} from './transform-types';
+import type {TransformProps, TransformDrawOptions, TransformRunOptions} from './transform';
 import Buffer, {BufferProps} from '../classic/buffer';
 import TransformFeedback from '../classic/transform-feedback';
 import Model from '../engine/classic-model';
